bm25_corpus_index/requirements.txt
google-cloud-storage
markdownify
pinecone-text
kfp
google-cloud-aiplatform
docker